+++ packages/svelte/tests/snapshot/samples/purity/_expected/client/index.svelte.js
@@ -8,11 +8,13 @@ export default function Purity($$anchor) {
 	var fragment = root();
 	var p = $.first_child(fragment);
 
-	p.textContent = '0';
+	p.textContent = (
+		$.untrack(() => Math.max(0, Math.min(0, 100)))
+	);
 
 	var p_1 = $.sibling(p, 2);
 
-	p_1.textContent = location.href;
+	p_1.textContent = ($.untrack(() => location.href));
 
 	var node = $.sibling(p_1, 2);